+++ tests/testthat/test-params.R
@@ -79,6 +79,7 @@ test_that("param ranges", {
   expect_equal(num_clusters(c(1, 2))$range, list(lower = 1, upper = 2))
   expect_equal(num_leaves(c(31,100))$range, list(lower = 31, upper = 100))
   expect_equal(trim_amount(c(0.1, 0.4))$range, list(lower = 0.1, upper = 0.4))
+  expect_equal(num_runs(c(2, 100))$range, list(lower = 2, upper = 100))
 })